requirements.txt
certifi==2020.6.20
Django==2.2.2
django-crispy-forms==1.7.2
idna==2.9
mysql-connector==2.2.9
mysqlclient==1.4.2.post1
psutil==5.7.2
PyMySQL==0.10.0
python-scrapyd-api==2.0.0
pytz==2020.1
requests==2.24.0
Scrapy==1.6.0
scrapy-djangoitem==1.1.1
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==16.6.0
tabulate==0.8.7